Backend/book_hub/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.5.0
gunicorn==21.2.0
dj-database-url==2.1.0 